+++ packages/uniapp-request/src/types/request.ts
@@ -77,6 +77,11 @@ export interface RequestOptions {
      * + `1.8.0` 及以上版本支持
      */
     originalResponse?: boolean;
+    /**
+     * 发起请求时的自定义参数，一般在拦截器里面使用
+     * + `v1.8.3` 及以上版本支持
+     */
+    customData?: any;
 };
 
 /**