+++ packages/uniapp-request/src/utils/constant.ts
@@ -12,3 +12,9 @@ export const NETWORK_TIMEOUT_CODE = 408;
  * 请求链接未找到状态码
  */
 export const URL_NOT_FOUND_CODE = 404;
+
+/**
+ * 请求失败统一状态码
+ * + `v1.8.3` 及以上版本支持
+ */
+export const NETWORK_FAILURE_CODE = -1;