+++ docs/changelog.md
@@ -1,6 +1,10 @@
 # 更新日志
 ---
 
+::: timeline 1.8.1	(2024-01-11)
++ 优化 `loading` 为 `false` 时微信开发者控制台报错问题。
+:::
+
 ::: timeline 1.8.0  (2023-08-11)
 + 新增 `originalResponse` 请求配置项。[详情](/config/request.html#originalresponse)
 :::